Remove offset in taskqueue search query Remove offset in taskqueue search query The offset in the query could cause runnable tasks to be ignored in certain cases. Offset should not be required because unrunnable tasks already do not get chosen since we ignore the dep hashes.
This commit also changes the hash type of dep hashes to md5 instead of
the hash builtin. Using the builtin was a mistake as it uses a seed
that changes each Python restart, resulting in a different hash for the
same data. This defeats the purpose of the dep hashes.